requirements.txt
# Core scientific stack
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Configuration
pyyaml>=6.0

# Dashboard
streamlit>=1.28.0
plotly>=5.17.0
altair>=5.0.0

# Notebook + visualization
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Console formatting
colorama>=0.4.6
